+++ CHANGELOG.md
@@ -6,9 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## 2.2.3 - 2020-10-22
 
+### Added
+
 ### Fixed
 
-* Fixed hardcoded subscription id
+* Fixed SubscriptionCallbackController not reseting