+++ CHANGELOG.md
@@ -4,13 +4,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 2.2.3 - 2020-10-22
-
-### Added
+## 2.2.5 - 2020-10-22
 
 ### Fixed
 
-* Fixed SubscriptionCallbackController not reseting
+* Fixed `QD\commerce\quickpay\elements\db\SubscriptionQuery` parsing `cardExpireYear` and `cardExpireMonth` as date instead of string
+
+## 2.2.4 - 2020-10-22
+
+### Added
+
+* Added `cardExpireYear` and `cardExpireMonth` to `QD\commerce\quickpay\elements\db\SubscriptionQuery`
+
+
+## 2.2.3 - 2020-10-22
+
+### Fixed
+
+* Fixed hardcoded subscription id